refactor(runtime): consolidate flat runtime_* modules into subpackages
- Move the ten bootstrap/runtime_*.py modules into a bootstrap/runtime/
  package (generations, reload, background, shutdown, channel_barrier,
  dispatcher, events, construction, inspection, memory), git-mv so
  history follows.
- Move run_cleanup_steps to core/common/cleanup.py: it is a generic
  lifecycle utility used by desktop_bridge and main.py, not a bootstrap
  concern; this also drops a desktop_bridge -> bootstrap utility import.
- Move the four desktop_bridge/runtime_*.py modules into a
  desktop_bridge/runtime/ package (service, apply, factory, role_tasks).
- Mirror the test tree (tests/backend/bootstrap/runtime/,
  tests/backend/desktop_bridge/runtime/, core/common/test_cleanup.py).
- Both package __init__ files deliberately re-export nothing to keep
  import cycles impossible between the reload path and bootstrap.tools.
