+++ apps/backend/desktop_bridge/runtime/__init__.py
@@ -1,0 +1,5 @@
+"""Bridge-side runtime reload surface: reloadable endpoint, settings apply, assembly.
+
+Modules are imported directly (``desktop_bridge.runtime.service`` etc.);
+this package re-exports nothing.
+"""
